+++ src/misc/tokens.ts
@@ -2,6 +2,8 @@ import {InjectionToken} from '@angular/core';
 import {Notifications} from '@anglr/common';
 
 import {HttpClientErrorResponseMapper, HttpClientValidationErrorResponseMapper} from './httpError.interface';
+import {AnglrExceptionExtender} from '../exceptionHandling/anglrExceptionExtender';
+import {DummyInternalServerErrorRenderer, InternalServerErrorRenderer} from '../internalServerError/internalServerError.service';
 
 /**
  * Injection token used for injecting notifications service used withing error handling package
@@ -22,3 +24,13 @@ export const CLIENT_ERROR_NOTIFICATIONS: InjectionToken<Notifications> = new Inj
  * Injection token that represents default response mapper for http client validation errors
  */
 export const HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER: InjectionToken<HttpClientValidationErrorResponseMapper> = new InjectionToken<HttpClientValidationErrorResponseMapper>('HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER');
+
+/**
+ * Injection token used for injecting multiple AnglrExceptionExtender 
+ */
+export const ANGLR_EXCEPTION_EXTENDERS: InjectionToken<AnglrExceptionExtender[]> = new InjectionToken<AnglrExceptionExtender[]>('ANGLR_EXCEPTION_EXTENDERS');
+
+/**
+ * Injection token used for InternalServerErrorRenderer
+ */
+export const INTERNAL_SERVER_ERROR_RENDERER: InjectionToken<InternalServerErrorRenderer> = new InjectionToken<InternalServerErrorRenderer>('INTERNAL_SERVER_ERROR_RENDERER', {providedIn: 'root', factory: () => new DummyInternalServerErrorRenderer()});